+++ src/Dap.Remote/RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+## 0.6.1
+* Improve logging with HubSpec
+
 ## 0.6.0
 * Use Thoth for json encode and decode
 * Type definition for remote access